refactor(queue): streamline queue registration process and improve imports
- Remove redundant queue registration lines in AlertQueueService
- Integrate BullBoard instance creation within QueueModule
- Simplify dependency handling in WebhooksService and related modules
