+++ services/workflows-service/src/common/queue/types.ts
@@ -7,3 +7,14 @@ export interface BullBoardInjectedInstance {
   boardInstance: ReturnType<typeof createBullBoard>;
   serverAdapter: ExpressAdapter;
 }
+
+export const bullBoardProvider = {
+  provide: BULLBOARD_INSTANCE_INJECTION_TOKEN,
+  useFactory: (): BullBoardInjectedInstance => {
+    const serverAdapter = new ExpressAdapter();
+    serverAdapter.setBasePath('/api/queues');
+    const boardInstance = createBullBoard({ queues: [], serverAdapter });
+
+    return { boardInstance, serverAdapter };
+  },
+};
