+++ services/workflows-service/src/webhooks/webhooks.module.ts
@@ -1,5 +1,3 @@
-import { createBullBoard } from '@bull-board/api';
-import { ExpressAdapter } from '@bull-board/express';
 import { HttpModule } from '@nestjs/axios';
 import { Inject, MiddlewareConsumer, Module } from '@nestjs/common';
 
@@ -12,21 +10,8 @@ import { WebhooksService } from './webhooks.service';
 
 @Module({
   imports: [AppLoggerModule, HttpModule, QueueModule, MonitoringModule],
-  providers: [
-    WebhooksService,
-    {
-      provide: BULLBOARD_INSTANCE_INJECTION_TOKEN,
-      useFactory: (): BullBoardInjectedInstance => {
-        const serverAdapter = new ExpressAdapter();
-        serverAdapter.setBasePath('/api/queues');
-
-        const boardInstance = createBullBoard({ queues: [], serverAdapter });
-
-        return { boardInstance, serverAdapter };
-      },
-    },
-  ],
-  exports: [WebhooksService, BULLBOARD_INSTANCE_INJECTION_TOKEN],
+  providers: [WebhooksService],
+  exports: [WebhooksService],
 })
 export class WebhooksModule {
   constructor(
